% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForesPrediction.R
\name{getPred.rf.finalHumidity}
\alias{getPred.rf.finalHumidity}
\title{Random forest prediction}
\usage{
getPred.rf.finalHumidity(
  climaticSystem,
  innerTemp0,
  innerTempF,
  innerHum0,
  extTemp0,
  extTempF,
  extHum0,
  extHumF,
  extRad0,
  extRadF,
  secondsON
)
}
\arguments{
\item{climaticSystem}{Numerical vector containing number 1 when the two fans and the recirculators are on, 2 when the three fans and the recirculators are on or 3 when the three fans, the recirculators and Cooling are on.}

\item{innerTemp0}{Numerical vector containing the temperature (in Celsius) inside the greenhouse at the moment of switching on the climatic systems.}

\item{innerTempF}{Numerical vector containing the temperature (in Celsius) inside the greenhouse at the moment of switching off the climatic systems.}

\item{innerHum0}{Numerical vector containing the relative humidity (percentage) inside the greenhouse at the moment of switching on the climatic systems.}

\item{extTemp0}{Numerical vector containing the temperature (in Celsius) outside the greenhouse at the moment of switching on the climatic systems.}

\item{extTempF}{Numerical vector containing the temperature (in Celsius) outside the greenhouse at the moment of switching off the climatic systems.}

\item{extHum0}{Numerical vector containing the relative humidity (percentage) outside the greenhouse at the moment of switching on the climatic systems.}

\item{extHumF}{Numerical vector containing the relative humidity (percentage) outside the greenhouse at the moment of switching off the climatic systems.}

\item{extRad0}{Numerical vector containing the radiation (in W/m^2) outside the greenhouse at the moment of switching on the climatic systems.}

\item{extRadF}{Numerical vector containing the radiation (in W/m^2) outside the greenhouse at the moment of switching off the climatic systems.}

\item{secondsON}{Numerical vector containing the switch-on times of the climatic systems in seconds.}
}
\value{
Numeric: indicating the final humidity inside the greenhouse.
}
\description{
Prediction of the final humidity inside the greenhouse, taking into account the different parameters.
}
\examples{
getPred.rf.finalHumidity(3, 27.4, 25.2, 70.5, 22.2, 21.7, 100, 98, 414, 477, 1980)

}
\author{
Jose Mendoza Bernal
}
